Acquire the dependencies of the package on any depth level.

This function is an ultimate wrapper for get_dependencies. It inherits all of the arguments and allows to recursively search for the dependencies at the higher level of depth.

deepdep(package, depth = 1, downloads = FALSE, bioc = FALSE,
  local = FALSE, deps_types = c("Depends", "Imports"))

Arguments

package
A character. Name of the package that is on CRAN, Bioconductor repository or locally installed. See bioc and local arguments.

depth
An integer. Maximum depth level of the dependency. By default it's 1.

downloads
A logical. If TRUE add dependency downloads data. By default it's FALSE.

bioc
A logical value. If TRUE the Bioconductor dependencies data will be taken from. Bioconductor repository. For this option to work properly, BiocManager package needs to be installed.

local
A logical value. If TRUE only data of locally installed packages will be used (without API usage).

deps_types
A character vector. Types of dependencies that should be sought. Possibilities are: "Imports", "Depends", "Suggests", "Enhances", "LinkingTo". By default it's "Depends","Imports".

Value

An object of deepdep class.

See also

Examples

library(deepdep)

dd <- deepdep("stringr", downloads = TRUE)
dd

#>           origin     name  version    type last_day last_week last_month
#> imports1 stringr     glue >= 1.2.0 Imports    15676    183834     789981
#> imports2 stringr magrittr     <NA> Imports   113881    840456    3430942
#> imports3 stringr  stringi >= 1.1.7 Imports    14818    164986     693655
#>          last_quarter last_half grand_total
#> imports1      2372075   4431625    14326922
#> imports2     10297405  15821253    30093579
#> imports3      2082935   3642208    21420843

dd2 <- deepdep("stringr", depth = 2)
dd2

#>           origin     name  version    type
#> imports1 stringr     glue >= 1.2.0 Imports
#> imports2 stringr magrittr     <NA> Imports
#> imports3 stringr  stringi >= 1.1.7 Imports

dd3 <- deepdep("ggplot2", local = TRUE)

#> Error in check_package_name(package, bioc, local): ggplot2 is not installed

dd3

#> Error in eval(expr, envir, enclos): nie znaleziono obiektu 'dd3'